[Move] TypeScript types
